keg_relocate: port to generic OS.
