+++ Library/Homebrew/extend/os/keg_relocate.rb
@@ -1,0 +1,5 @@
+require "keg_relocate"
+
+if OS.mac?
+  require "extend/os/mac/keg_relocate"
+end
